Testing TDC xml file import
Here I'm testing importing an xml file! Currently it only connects to the 'mode set' dialog's checkboxes, but it's just to see how the xml data importing process works. At the moment all the values set in the xml file are the same as the TDCreg file, but when I change them in the xml file and press 'load setup', then they are changed in the GUI to the xml values!