creating/reading xml file
Hi Yuxiang! Here, the file 'create_TDC_default.py' creates the xml file 'TDC_default.xml'. The values set in the xml file are loaded into the GUI when the 'load setup' button is pressed! I put in variables different from the ones set in TDCreg.py, so that the changes can be seen in the GUI. So far, when the init button is pressed, then all the new values are run!